requirements.txt
langchain
langchain-huggingface
langchain-text-splitters
langchain-community
langchain-chroma
chromadb
sentence-transformers
pypdf
gradio
python-dotenv
huggingface-hub
langgraph